Add date time picker E2E
